refactor: Added authControllers
All auth functions replaced to auth-folder
